backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.1
pydantic==2.6.4
pydantic-settings==2.2.1
sqlalchemy==2.0.28
supabase==2.3.0
langchain==0.1.13
langchain-community==0.0.29
langchain-core==0.1.33
langchain-groq==0.0.1
langchain-google-genai==0.0.11
google-auth
google-auth-httplib2
google-auth-oauthlib
requests==2.31.0
slowapi==0.1.9
twilio==9.0.4
stripe==8.6.0
python-multipart==0.0.9
email-validator
# New Additions for Async Workers & AI Migration
redis==5.0.3
hiredis==2.3.2
arq==0.26.0
phidata==2.4.21
openai==1.14.3
razorpay==1.3.0
